requirements.txt
dm-haiku==0.0.11
optax==0.1.7
numpy==1.22.3
tqdm==4.64.0
datasets==2.0.0
Pillow==9.3.0
black
